Document and rename loadEnv
